+++ packages/banner/src/Banner.story.tsx
@@ -2,14 +2,22 @@ import React from 'react';
 import Icon, { glyphs } from '@leafygreen-ui/icon';
 import Banner, { Variant } from '.';
 import { ComponentStory, Meta } from '@storybook/react';
-import { storybookArgTypes } from '@leafygreen-ui/lib/';;
+import { storybookArgTypes } from '@leafygreen-ui/lib/';
+import { Link } from '@leafygreen-ui/typography';
 
 export default {
   title: 'Components/Banner',
   component: Banner,
+  parameters: {
+    controls: {
+      exclude: ['ref', 'className', 'onClose'],
+    },
+  },
   args: {
     children:
       'To avoid disrupting majority writes, new members are now added to replica sets as priority=0, votes=0 until they reach secondary state, after which Cloud Manager automatically updates the configuration to match the priority and votes value specified in the deployment.',
+    darkMode: false,
+    dismissible: false,
   },
   argTypes: {
     image: {
@@ -24,6 +32,7 @@ export default {
     },
     children: storybookArgTypes.children,
     ref: storybookArgTypes.ref,
+    darkMode: storybookArgTypes.darkMode,
   },
 } as Meta<typeof Banner>;
 
@@ -54,6 +63,28 @@ Danger.args = {
 export const Info = Template.bind({});
 Info.args = {
   variant: Variant.Info,
+};
+
+export const WithLink: ComponentStory<typeof Banner> = ({
+  // eslint-disable-next-line react/prop-types
+  image,
+  // eslint-disable-next-line react/prop-types
+  children,
+  ...args
+}) => {
+  return (
+    <Banner
+      // @ts-expect-error
+      image={image ? <Icon glyph={image} /> : undefined}
+      {...args}
+    >
+      {children}
+      &nbsp;
+      <Link href="http://localhost:9001">Link style</Link>
+      &nbsp;
+      <a href="http://localhost:9001">Regular link</a>
+    </Banner>
+  );
 };
 
 export const WithCustomImage = Template.bind({});